removed Cards.ios
