--- packages/palette/src/elements/Cards/LargeCard.shared.tsx
@@ -1,8 +1,0 @@
-import { CardTagProps } from "./CardTag.shared"
-
-export interface LargeCardProps {
-  image: string
-  title: string
-  subtitle?: string
-  tag?: CardTagProps
-}
