+++ packages/palette/src/elements/Cards/LargeCard.tsx
@@ -1,5 +1,12 @@
 import React from "react"
-import { LargeCardProps } from "./LargeCard.shared"
+import { CardTagProps } from "./CardTag"
+
+export interface LargeCardProps {
+  image: string
+  title: string
+  subtitle?: string
+  tag?: CardTagProps
+}
 
 /**
  * `Large` is a card with one image one tall image, and text for title and subtitle
