--- packages/palette/src/elements/Cards/MediumCard.shared.tsx
@@ -1,8 +1,0 @@
-import { CardTagProps } from "./CardTag.shared"
-
-export interface MediumCardProps {
-  image: string
-  title: string
-  subtitle?: string
-  tag?: CardTagProps
-}
