--- packages/palette/src/elements/Cards/SmallCard.shared.tsx
@@ -1,8 +1,0 @@
-import { CardTagProps } from "./CardTag.shared"
-
-export interface SmallCardProps {
-  images: string[]
-  title: string
-  subtitle?: string
-  tag?: CardTagProps
-}
