+++ packages/palette/src/elements/Cards/index.tsx
@@ -1,8 +1,4 @@
 export * from "./SmallCard"
-export * from "./SmallCard.shared"
 export * from "./MediumCard"
-export * from "./MediumCard.shared"
 export * from "./LargeCard"
-export * from "./LargeCard.shared"
 export * from "./CardTag"
-export * from "./CardTag.shared"
